Cleanup pass: workspace builds clean, types consolidated, half-built crates rescued (#3514)
* cleanup: tier A — slop removal + minor fixes

- Delete 7 section banner triples in scan_engine.rs + llm-client integration tests
- Drop SpawnAlertConfig/Stats/Page/Entry/WatchPattern duplicate block in web/src/types.ts (lines 975-1010)
- Type personalBests as PersonalBest[] in SessionDetail.tsx
- SDK client.rs: replace 8x .unwrap_or_default() on error body with explicit fallback message
- Delete textquest-soul/src/debrief_aggregator.rs (unimplemented stub, no callers, broken Eq derive)
- transport.rs: 6 TODO comments → FIXME with concrete pending wiring
- EconomyPanel.tsx: delete 6 TODO comments on local-state-only handlers

Skipped (agent over-flag):
- A2 windows-dangerous-tests (cfg-gate is live in textquest-dll/src/stealth/thread_pool.rs)
- A2 runtime-only feature (used by textquest-dll dependency edge)
- A9 redundant rustdoc (mostly legitimate WHY/protocol docs)

Deferred (textquest-learn rescue, not cleanup):
- A3-A6 (broken PolicyMetadata/PolicyManifest re-exports, missing tracing dep, broken cfg(feature = full) gates)

* cleanup(b1): consolidate ErrorResponse to textquest-common::api_types

Single canonical `pub struct ErrorResponse { pub error: String }` with serde derives
in textquest-common. Replaces 7 internal duplicates:

- textquest-web/src/api/control.rs
- textquest-web/src/api/inventory_utility_parity.rs
- textquest-web/src/api/xassist.rs
- textquest-web/src/api/admin_sessions.rs
- textquest-web/src/api.rs
- textquest-web-sdk/src/models.rs (now re-exported)
- textquest/src/bin/admin_client/mod.rs

Skipped:
- textquest-client/rust/src/models.rs — separate distribution boundary, no common dep
- TS ErrorResponse in tools/typescript-sdk + textquest-client/typescript — handled separately

Risk: low. All Rust defs were wire-identical { error: String }.

* cleanup(b5): consolidate json_error helper to textquest-web/src/error.rs

Single helper module with two variants:
- json_error(status, msg) -> Response (for handlers returning Response)
- json_error_pair(status, msg) -> (StatusCode, Json<ErrorResponse>) (tuple form)

Replaces 10 local copies (4 tuple + 6 Response). Removes 4 stray local
ErrorResponse struct definitions in api/auto_camp.rs, api/auto_rez.rs,
api/auto_group.rs, api/paranoid.rs that B1 didn't catch (they were private).

extensions.rs json_error used hand-rolled json!({"error": ...}) — output
JSON shape identical to ErrorResponse, so swap is wire-compatible.

* fix: textquest-learn + textquest-dll buildable on stable

textquest-learn:
- Add tracing/tracing-subscriber/clap/tokio/rusqlite/ndarray deps to Cargo.toml
- Declare 'full' feature; bin/main.rs uses required-features = ['full']
- Define PolicyMetadata struct in policy.rs from usage callsites (8 fields)
- Define OpeReport + extended PolicyManifest with ope_report field
- Add ExperienceEntry.context: Value field + #[serde(default)] on optional fields
- Add ExperienceLedger::load_from_dir alias
- Fix get_segment(start_idx, end_idx) signature: usize -> u64 to match Bookmark
- Fix reward_golden_test clamp ambiguity (f32 type annotation)
- Replace dead smoke.rs trait imports with current public types

textquest-dll:
- Declare combat::bandolier in mod.rs (file existed, decl was missing)
- Replace #[derive(Debug)] with manual Debug impl on EventWriter (zstd::Encoder lacks Debug)
- Cast i32 -> i64 at clamp_to_u32 callsites for mana_current/endurance_current
- textquest-common: add PositionalContext.is_stealthed field
- textquest-common: add ConditionExpr::PlayerIsStealthed variant
- Bulk-add burn_state/burnnow_triggered/burn_cooldown_ticks (94 sites) to test CombatContext literals
- Bulk-add burn_duration_ticks/burn_cooldown_duration_ticks (13 sites) to test RotationGroup literals
- Add positional: None to 6 strategy/beastlord CombatContext literals
- Add ConditionExpr::PlayerIsStealthed match arm in evaluate_condition

* fix: workspace + all tests cargo check clean

textquest-soul lancedb/arrow API migration:
- Float32Type moved to arrow_array::types
- Add lancedb::query::{ExecutableQuery, QueryBase} for only_if method
- DataType::FixedSizeList + FixedSizeListArray::from_iter_primitive: dim arg usize → i32
- table.add: explicit Box<dyn RecordBatchReader + Send> coercion

textquest-soul/llm/fallback test:
- respond_to_player call: add missing 5th memory_context arg

textquest replay.rs:
- entry.path() into_owned() so subsequent unpack_in() can take &mut
- Add use std::io::BufRead in test scope for Vec<u8>::lines()

textquest config.rs:
- Remove duplicate `pub improve` field declaration

textquest stats:
- Gate test_stats_aggregation.rs with #![cfg(windows)] (matches lib gate)

textquest class_config:
- Add modes/active_mode/burn_rotation to test fixtures

textquest-learn:
- Gate bc_integration_test with #![cfg(feature = "full")]
- Use policy::PolicyMetadata via explicit path

textquest-web:
- Fix migrations include_str path: ../../ → ../../../ (3 levels from sessions_db/)
- migrations.rs: use unchecked_transaction (we hold &Connection intentionally)
- accounts.rs: import ErrorResponse from textquest_common::api_types
- Replace 3 ad-hoc AppState fixtures with crate::test_support::demo_app_state()
- test_support: add raid_config, sound_config, text_to_speech_state, self_improvement_state, suggestion_state, config_change_history, last_config_change
- Add demo_app_state_with_auth helper for auth-enabled tests
- main.rs test_state_with_credentials: add 5 missing AppState fields
- alerts.rs: add audio to AlertingConfig literal
- api.rs: add improve_auto_promote to CharacterConfigUpdate literal

llm-client:
- Delete dead Ollama integration test (Ollama backend removed; current API is LlmBackend trait)

textquest-common:
- combat: add PositionalContext.is_stealthed
- combat: add ConditionExpr::PlayerIsStealthed variant

* cleanup: clippy autofix + tmp removal + bandolier first-swap fix + Tier C/D1

- cargo clippy --workspace --tests --fix: ~50+ lints autofixed across 25 files (unused imports, redundant clones, dead code, format string args, & on slice idioms)
- Delete 14 .rs.tmp orphan files in textquest-dll/src/combat/classes/ (frozen at bbed2ef6 burnnow rotation commit; .rs counterparts evolved past them)
- Fix BandolierManager first-swap cooldown bug: last_swap_tick=0 now sentinel for 'never swapped'; activate_command and evaluate_rules skip cooldown gate on first swap (5 tests now pass: activate_command_returns_correct_format, _escapes_quotes_and_backslashes, rule_engine_first_rule_wins, _fires_on_tank_aggro, _distance_above_triggers)
- Trim 'legacy provenance reporting' from inventory_utility_parity doc comment (D2 — comment-only legacy reference, no actual code path)
- useBackendStatus.ts: console.warn on health-check fetch failure so operators can debug network errors (C1)

D1 LegacyProfile / D3 ETW 26-byte / D4 spell-system: verified live in code paths, kept.

+++ textquest/src/travel/graph.rs
@@ -663,7 +663,7 @@ mod tests {
             zone_points.push(point(
                 &format!("z{index}"),
                 "LiveEra",
-                index as i32,
+                index,
                 0,
                 0,
             ));
